feat: support non-stdin formatters
Fixes #1

For formatters that don't support stdin you can pass a `tempfile` field which
can be either `"random"` (uses `vim.fn.tempname()`) or a function that returns a
string to be used as the tempfile and then the plugin will write the buffer to
this file, run the command on it (the "%" value will be expanded to the
tempfile) and read it back and fill in the buffer).

I added an example to the README file demonstrating how to use goimports-reviser.