fix: preserve scroll when restoring cursor position in other windows
Previously restoring cursor position in other windows in the same tab didn't
work well because it would still scroll to the top (if you just switch to that
window it would automatically scroll back to the old position).

I'm not sure why that happens but for now instead of calling the restore
function directly I wrapped it with `vim.fn.timer_start` (with a delay of 0) and
then it works (there's an annoying flicker but at least it doesn't stay at the
top).

I hope I can figure out a better way to solve this someday but for now it's a
big improvement (IMO).